fix: controllers: album, donation, event, message and news

+++ src/controllers/albumsController.js
@@ -33,9 +33,13 @@ exports.createAlbums = (req, res) => {
     });
     if (req.file && req.file.filename) {
       newAlbum.imgURL = `${req.file.filename}`;
-      newAlbum.save();
-      res.send({ message: "registro de foto correctamente" });
+
+    }else{
+      newAlbum.imgURL = "ifgf.png";
     }
+
+    newAlbum.save();
+    res.send({ message: "registro de foto correctamente" });
   } catch (error) {
     res.send({ message: error });
   }
@@ -80,7 +84,9 @@ exports.deleteAlbumsById = async (req, res) => {
   try {
     const id = req.params.id;
     const doc = await model.findOneAndDelete({ _id: parseId(id) });
-    unlink(path.resolve("./uploads/" + doc.imgURL));
+    if(!doc.imgURL === "ifgf.png"){
+      unlink(path.resolve("./uploads/" + doc.imgURL));
+    }
     res.send({ message: "Eliminado con exito" });
   } catch (error) {
     res.send({ message: error });
